+++ modules/aws-anyscale-iam/test/anyscale-test/main.tf
@@ -34,6 +34,23 @@ module "iam_cluster_node_instance_profile" {
 }
 
 # ---------------------------------------------------------------------------------------------------------------------
+# Create the cluster node and instance profile for Anyscale with no optional parameters. No access role.
+# ---------------------------------------------------------------------------------------------------------------------
+module "iam_secretsmanager_instance_profile" {
+  source = "../.."
+
+  module_enabled                       = true
+  create_anyscale_access_role          = false
+  create_cluster_node_instance_profile = true
+  anyscale_cluster_node_role_name      = "anyscale-cluster-node-role-secretsmanager"
+  anyscale_cluster_node_byod_secret_arns = [
+    "arn:aws:secretsmanager:us-west-2:367974485317:secret:brent/tf/byod-test-kms_encrypted-HE5rlP"
+  ]
+  anyscale_cluster_node_byod_secret_kms_arn = "arn:aws:kms:us-west-2:367974485317:key/c11ad123-4353-42e2-99c9-2c2cf9b4d943"
+  tags                                      = local.full_tags
+}
+
+# ---------------------------------------------------------------------------------------------------------------------
 # Use all params and build both roles.
 # ---------------------------------------------------------------------------------------------------------------------
 module "kitchen_sink" {
@@ -56,7 +73,6 @@ module "kitchen_sink" {
   anyscale_access_servicesv2_policy_path        = "/testpath/"
   anyscale_access_servicesv2_policy_description = "Anyscale Services v2 TESTPOLICY Description"
 
-  create_iam_s3_policy               = true
   anyscale_s3_bucket_arn             = "arn:aws:s3:::anyscale-demo"
   anyscale_iam_s3_policy_name_prefix = "anyscale-s3-testpolicy-"
   anyscale_iam_s3_policy_path        = "/tests3path/"
@@ -83,6 +99,10 @@ module "kitchen_sink" {
 
   anyscale_cluster_node_managed_policy_arns = [
     "arn:aws:iam::aws:policy/AmazonGlacierReadOnlyAccess"
+  ]
+
+  anyscale_cluster_node_byod_secret_arns = [
+    "arn:aws:secretsmanager:us-west-2:367974485317:secret:brent/example-G3lNwD"
   ]
 
   tags = local.full_tags